move Block data generator test helper to test support (#391)
