+++ ethereum/eth/src/test/java/tech/pegasys/pantheon/ethereum/eth/messages/NodeDataMessageTest.java
@@ -12,9 +12,9 @@
  */
 package tech.pegasys.pantheon.ethereum.eth.messages;
 
+import tech.pegasys.pantheon.ethereum.core.BlockDataGenerator;
 import tech.pegasys.pantheon.ethereum.p2p.api.MessageData;
 import tech.pegasys.pantheon.ethereum.p2p.wire.RawMessage;
-import tech.pegasys.pantheon.ethereum.testutil.BlockDataGenerator;
 import tech.pegasys.pantheon.util.bytes.BytesValue;
 
 import java.util.ArrayList;
