+++ ethereum/eth/src/test/java/tech/pegasys/pantheon/ethereum/eth/messages/ReceiptsMessageTest.java
@@ -12,10 +12,10 @@
  */
 package tech.pegasys.pantheon.ethereum.eth.messages;
 
+import tech.pegasys.pantheon.ethereum.core.BlockDataGenerator;
 import tech.pegasys.pantheon.ethereum.core.TransactionReceipt;
 import tech.pegasys.pantheon.ethereum.p2p.api.MessageData;
 import tech.pegasys.pantheon.ethereum.p2p.wire.RawMessage;
-import tech.pegasys.pantheon.ethereum.testutil.BlockDataGenerator;
 
 import java.util.ArrayList;
 import java.util.Iterator;
